snes.inc: add symbols for high-byte
In #3, a user of the template reported confusion between reading
the controller's entire report as a 16-bit unit and reading only
the high byte as an 8-bit unit.  To make this confusion less
likely going forward, add a set of labels and constants for
reading the high byte for convenience in porting NES code.